packages/wafer_fault_detection/requirements.txt
appdirs==1.4.4
attrs==20.3.0
click==7.1.2
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
imbalanced-learn==0.8.0
importlib-metadata==3.7.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
kneed==0.7.0
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-mock==3.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
tox==3.23.0
typing-extensions==3.7.4.3
virtualenv==20.4.2
Werkzeug==1.0.1
xgboost==1.3.3
zipp==3.4.1
